Refactor catalog further
